streamlit/requirements.txt
numpy
Pillow
streamlit
requests==2.28.1
